Reorganizar actions y components. Usar isomorphic-fetch.
