+++ src/components/App.js
@@ -1,7 +1,7 @@
-import React from 'react'
-import Footer from './Footer'
-import NavbarUTNianos from '../containers/NavbarUTNianos'
-import Main from '../containers/Programa'
+import React from 'react';
+import Footer from './Footer';
+import NavbarUTNianos from '../containers/NavbarUTNianos';
+import Main from '../containers/Main';
 
 const App = () => (
   <div>
@@ -9,6 +9,6 @@ const App = () => (
     <Main />
     <Footer />
   </div>
-)
+);
 
-export default App
+export default App;
